+++ app/incometax/unauthorisedagent/connectors/SubscriptionStoreConnector.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package incometax.subscription.connectors
+package incometax.unauthorisedagent.connectors
 
 import javax.inject.Inject
 
 import core.config.AppConfig
 import incometax.subscription.httpparsers.DeleteSubscriptionResponseHttpParser._
 import incometax.subscription.httpparsers.RetrieveSubscriptionResponseHttpParser._
-import incometax.subscription.httpparsers.StoreSubscriptionResponseHttpParser._
+import incometax.unauthorisedagent.httpparsers.StoreSubscriptionResponseHttpParser._
 import incometax.subscription.models.StoredSubscription
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.bootstrap.http.HttpClient
