+++ app/incometax/unauthorisedagent/controllers/ConfirmAgentSubscriptionController.scala
@@ -14,17 +14,16 @@
  * limitations under the License.
  */
 
-package usermatching.controllers
+package incometax.unauthorisedagent.controllers
 
 import javax.inject.{Inject, Singleton}
 
 import core.ITSASessionKeys
-import core.services.KeystoreService
 import core.ITSASessionKeys._
 import core.auth.AuthenticatedController
 import core.config.BaseControllerConfig
-import core.services.AuthService
 import core.services.CacheUtil._
+import core.services.{AuthService, KeystoreService}
 import incometax.subscription.models.SubscriptionSuccess
 import incometax.subscription.services.SubscriptionOrchestrationService
 import play.api.i18n.MessagesApi
@@ -45,7 +44,7 @@ class ConfirmAgentSubscriptionController @Inject()(val baseConfig: BaseControlle
     implicit req =>
       user =>
         val arn = req.session(AgentReferenceNumber)
-        Ok(usermatching.views.html.confirm_agent_subscription(arn))
+        Ok(incometax.unauthorisedagent.views.html.confirm_agent_subscription(arn))
   }
 
   def submit(): Action[AnyContent] = Authenticated.async {
