+++ conf/app.routes
@@ -88,5 +88,5 @@ GET         /error/user-details           usermatching.controllers.UserDetailsEr
 POST        /error/user-details           usermatching.controllers.UserDetailsErrorController.submit()
 
 # Confirm agent subscription
-GET         /confirm-agent-subscription   usermatching.controllers.ConfirmAgentSubscriptionController.show()
-POST        /confirm-agent-subscription   usermatching.controllers.ConfirmAgentSubscriptionController.submit()
+GET         /confirm-agent-subscription   incometax.unauthorisedagent.controllers.ConfirmAgentSubscriptionController.show()
+POST        /confirm-agent-subscription   incometax.unauthorisedagent.controllers.ConfirmAgentSubscriptionController.submit()
