+++ test/incometax/unauthroisedagent/connectors/mocks/MockSubscriptionStoreConnector.scala
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package incometax.subscription.connectors.mocks
+package incometax.unauthroisedagent.connectors.mocks
 
 import core.utils.MockTrait
-import incometax.subscription.connectors.SubscriptionStoreConnector
+import core.utils.TestConstants._
+import core.utils.TestModels._
+import incometax.subscription.httpparsers.DeleteSubscriptionResponseHttpParser.DeleteSubscriptionResponse
 import incometax.subscription.httpparsers.RetrieveSubscriptionResponseHttpParser._
+import incometax.subscription.models.{DeleteSubscriptionFailure, DeleteSubscriptionSuccess, RetrieveSubscriptionFailure}
+import incometax.unauthorisedagent.connectors.SubscriptionStoreConnector
 import org.mockito.ArgumentMatchers
 import org.mockito.Mockito._
 import uk.gov.hmrc.http.HeaderCarrier
-import core.utils.TestModels._
-import core.utils.TestConstants._
-import incometax.subscription.httpparsers.DeleteSubscriptionResponseHttpParser.DeleteSubscriptionResponse
-import incometax.subscription.models.{DeleteSubscriptionFailure, DeleteSubscriptionSuccess, RetrieveSubscriptionFailure}
 
 import scala.concurrent.Future
