+++ test/incometax/unauthroisedagent/services/SubscriptionStoreServiceSpec.scala
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package incometax.subscription.services
+package incometax.unauthroisedagent.services
 
 import core.config.featureswitch.FeatureSwitching
 import core.utils.TestConstants._
 import core.utils.TestModels._
 import core.utils.UnitTestTrait
 import incometax.subscription.models.DeleteSubscriptionSuccess
-import incometax.subscription.services.mocks.TestSubscriptionStoreService
+import incometax.unauthroisedagent.services.mocks.TestSubscriptionStoreService
 import play.api.test.Helpers._
 import uk.gov.hmrc.http.InternalServerException
